fix issues #1 and #2
bug #1: When I try to parse the xml files in line 785 and 1011 I get an error
message that the '<' sign is not in line 1 column 1. When inspecting the
file with open(fp).readline() we discovered that the xml's contain a
Byte Order Mark. We have been able to get around this by adjusting line
784 and line 1010 as follows:
	with open(fp, 'r', encoding="utf-8") as fh:
ecoinvent now parses properly

--------
bug 2: missing parenthesis

 On branch characterisation
 Changes to be committed:
	modified:   ecospold2matrix.py