When positive_waste, add "[treatment for]" to waste labels
With positive_waste, we are changing the sign convention, treating waste
flows as positive, but that does not mean that we need a positive input
of waste. It only means that we require a positive amount of treatment
for waste... just a different convention. We want the labels to clearly
distinguish that.

 On branch dev
 Your branch is up-to-date with 'origin/dev'.

 Changes to be committed:
	modified:   ecospold2matrix.py